CAUSEWAY-3613: moves excel download classes into correct package
in line with other extensions
